+++ rust/operator-binary/src/service.rs
@@ -3,7 +3,7 @@ use stackable_operator::{
     builder::meta::ObjectMetaBuilder,
     commons::product_image_selection::ResolvedProductImage,
     k8s_openapi::api::core::v1::{Service, ServicePort, ServiceSpec},
-    kvp::{Label, Labels},
+    kvp::{Annotations, Labels},
     role_utils::RoleGroupRef,
 };
 
@@ -39,8 +39,7 @@ pub fn build_rolegroup_headless_service(
     let headless_service = Service {
         metadata: ObjectMetaBuilder::new()
             .name_and_namespace(hive)
-            // TODO: Use method on RoleGroupRef once op-rs is released
-            .name(rolegroup_headless_service_name(rolegroup))
+            .name(rolegroup.rolegroup_headless_service_name())
             .ownerreference_from_resource(hive, None, Some(true))
             .context(ObjectMissingMetadataForOwnerRefSnafu)?
             .with_recommended_labels(build_recommended_labels(
@@ -79,8 +78,7 @@ pub fn build_rolegroup_metrics_service(
     let metrics_service = Service {
         metadata: ObjectMetaBuilder::new()
             .name_and_namespace(hive)
-            // TODO: Use method on RoleGroupRef once op-rs is released
-            .name(rolegroup_metrics_service_name(rolegroup))
+            .name(rolegroup.rolegroup_metrics_service_name())
             .ownerreference_from_resource(hive, None, Some(true))
             .context(ObjectMissingMetadataForOwnerRefSnafu)?
             .with_recommended_labels(build_recommended_labels(
@@ -90,7 +88,8 @@ pub fn build_rolegroup_metrics_service(
                 &rolegroup.role_group,
             ))
             .context(MetadataBuildSnafu)?
-            .with_label(Label::try_from(("prometheus.io/scrape", "true")).context(LabelBuildSnafu)?)
+            .with_labels(prometheus_labels())
+            .with_annotations(prometheus_annotations())
             .build(),
         spec: Some(ServiceSpec {
             // Internal communication does not need to be exposed
@@ -110,18 +109,6 @@ pub fn build_rolegroup_metrics_service(
     Ok(metrics_service)
 }
 
-/// Headless service for cluster internal purposes only.
-// TODO: Move to operator-rs
-pub fn rolegroup_headless_service_name(rolegroup: &RoleGroupRef<v1alpha1::HiveCluster>) -> String {
-    format!("{name}-headless", name = rolegroup.object_name())
-}
-
-/// Headless metrics service exposes Prometheus endpoint only
-// TODO: Move to operator-rs
-pub fn rolegroup_metrics_service_name(rolegroup: &RoleGroupRef<v1alpha1::HiveCluster>) -> String {
-    format!("{name}-metrics", name = rolegroup.object_name())
-}
-
 fn metrics_ports() -> Vec<ServicePort> {
     vec![ServicePort {
         name: Some(METRICS_PORT_NAME.to_string()),
@@ -138,4 +125,24 @@ fn service_ports() -> Vec<ServicePort> {
         protocol: Some("TCP".to_string()),
         ..ServicePort::default()
     }]
+}
+
+/// Common labels for Prometheus
+fn prometheus_labels() -> Labels {
+    Labels::try_from([("prometheus.io/scrape", "true")]).expect("should be a valid label")
+}
+
+/// Common annotations for Prometheus
+///
+/// These annotations can be used in a ServiceMonitor.
+///
+/// see also <https://github.com/prometheus-community/helm-charts/blob/prometheus-27.32.0/charts/prometheus/values.yaml#L983-L1036>
+fn prometheus_annotations() -> Annotations {
+    Annotations::try_from([
+        ("prometheus.io/path".to_owned(), "/metrics".to_owned()),
+        ("prometheus.io/port".to_owned(), METRICS_PORT.to_string()),
+        ("prometheus.io/scheme".to_owned(), "http".to_owned()),
+        ("prometheus.io/scrape".to_owned(), "true".to_owned()),
+    ])
+    .expect("should be valid annotations")
 }